+++ .storybook/preview.ts
@@ -47,7 +47,7 @@ const preview: Preview = {
       }
     },
     a11y: {
-      test: 'todo'
+      test: 'error'
     }
   }
 };